+++ packages/cli/src/playground/src/domains/agents/agent-information.tsx
@@ -1,19 +1,22 @@
-import { useNavigate } from 'react-router';
-
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs';
 
 import { AgentDetails } from './agent-details';
 import { AgentEndpoints } from './agent-endpoints';
+import { AgentPromptEnhancer } from './agent-instructions-enhancer';
 import { AgentLogs } from './agent-logs';
 
 export function AgentInformation({ agentId }: { agentId: string }) {
-  const navigate = useNavigate();
   return (
     <Tabs defaultValue="details">
       <TabsList className="flex shrink-0 border-b">
         <TabsTrigger value="details" className="group">
           <p className="text-xs p-3 text-mastra-el-3 group-data-[state=active]:text-mastra-el-5 group-data-[state=active]:border-b-2 group-data-[state=active]:pb-2.5 border-white">
             Details
+          </p>
+        </TabsTrigger>
+        <TabsTrigger value="instructions" className="group">
+          <p className="text-xs p-3 text-mastra-el-3 group-data-[state=active]:text-mastra-el-5 group-data-[state=active]:border-b-2 group-data-[state=active]:pb-2.5 border-white">
+            Instructions
           </p>
         </TabsTrigger>
         <TabsTrigger value="endpoints" className="group">
@@ -26,9 +29,6 @@ export function AgentInformation({ agentId }: { agentId: string }) {
             Log Drains
           </p>
         </TabsTrigger>
-        <button className="group" onClick={() => navigate(`/agents/${agentId}/evals`)}>
-          <p className="text-xs p-3 text-mastra-el-3">Evals</p>
-        </button>
       </TabsList>
       <TabsContent value="details">{agentId ? <AgentDetails agentId={agentId} /> : null}</TabsContent>
       <TabsContent value="endpoints">
@@ -36,6 +36,9 @@ export function AgentInformation({ agentId }: { agentId: string }) {
       </TabsContent>
       <TabsContent value="logs">
         <AgentLogs agentId={agentId} />
+      </TabsContent>
+      <TabsContent value="instructions">
+        <AgentPromptEnhancer agentId={agentId} />
       </TabsContent>
     </Tabs>
   );