+++ packages/memory/integration-tests/src/with-libsql-storage.test.ts
@@ -1,5 +1,4 @@
-import { createOpenAI } from '@ai-sdk/openai';
-import { DefaultStorage, DefaultVectorDB } from '@mastra/core/storage';
+import { DefaultStorage } from '@mastra/core/storage';
 import { Memory } from '@mastra/memory';
 import dotenv from 'dotenv';
 import { describe } from 'vitest';
@@ -8,29 +7,37 @@ import { getResuableTests } from './reusable-tests';
 
 dotenv.config({ path: '.env.test' });
 
-const openai = createOpenAI({ apiKey: process.env.OPENAI_API_KEY });
-
 describe('Memory with LibSQL Integration', () => {
-  const vector = new DefaultVectorDB({
-    connectionUrl: 'file:test.db',
+  describe('with explicit storage', () => {
+    const memory = new Memory({
+      storage: new DefaultStorage({
+        config: {
+          url: 'file:test.db',
+        },
+      }),
+      options: {
+        lastMessages: 10,
+        semanticRecall: {
+          topK: 3,
+          messageRange: 2,
+        },
+      },
+    });
+
+    getResuableTests(memory);
   });
 
-  const memory = new Memory({
-    storage: new DefaultStorage({
-      config: {
-        url: 'file:test.db',
+  describe('with default storage', () => {
+    const memory = new Memory({
+      options: {
+        lastMessages: 10,
+        semanticRecall: {
+          topK: 3,
+          messageRange: 2,
+        },
       },
-    }),
-    vector,
-    options: {
-      lastMessages: 10,
-      semanticRecall: {
-        topK: 3,
-        messageRange: 2,
-      },
-    },
-    embedder: openai.embedding('text-embedding-3-small'),
-  });
+    });
 
-  getResuableTests(memory);
+    getResuableTests(memory);
+  });
 });